+++ Changelog.md
@@ -1,7 +1,9 @@
 # FiReS
 
 ## git master
+* [!19] Added casters for enum properties.
 * [!18] Fixed Visual Studio 14 compilation.
+* [!17] Added property gid getters in PropertyManager
 * [!16] Added property gid getters in PropertyManager
 * [!15] Added notifier/observer functionallity
 * [!14] Added property manager